requirements.txt
# pip install -U -r requirements.txt
Cython
numpy
opencv-python >= 4.1
torch >= 1.4
matplotlib
pycocotools
tqdm
Pillow
ffmpeg-python